feat(mojo): add aggregate mojo
Adding support for a new `restdocs-spec:aggregate` mojo that will allow
generation of a single spec file from across all modules in a
multi-module project.
